pyproject.toml
[project]
name = "repl-nix-aiteammatebuilders"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.53.0",
    "psycopg2-binary>=2.9.10",
    "langchain>=0.3.7",
    "langchain-community>=0.3.5",
    "langchain-openai>=0.2.5",
    "flask-cors>=5.0.0",
    "pygithub>=2.4.0",
    "requests>=2.32.3",
]
